ENTREGA  #4: Creada clase ShoppingItem. Creados RecyclerViewadapter, Viewholder, view_holder.xml Añadido RecyclerView a Mylistactivity. Cuando se inicia la actividad, o se clica en un item se añade, al recuclerView.

+++ app/src/main/java/com/example/ahorroplus/ShoppingItem.java
@@ -33,9 +33,9 @@ public class ShoppingItem {
             this.precioCarrefour=precioCarrefour;
             this.precioEroski=precioEroski;
 
-            int precioEros = Integer.parseInt(precioEroski);
-            int precioMerc = Integer.parseInt(precioMercadona);
-            int precioCarr = Integer.parseInt(precioCarrefour);
+            double precioEros = Double.parseDouble(precioEroski);
+            double precioMerc = Double.parseDouble(precioMercadona);
+            double precioCarr = Double.parseDouble(precioCarrefour);
 
             if(precioCarr<=precioEros && precioCarr<=precioMerc){
                 return precioCarrefour;
